@@ -1,28 +1,22 @@
-package co.ninetynine.android.exercise;
+package co.ninetynine.android.exercise
 
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.widget.TextView;
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import co.ninetynine.android.exercise.adapters.SectionsListAdapter
+import co.ninetynine.android.exercise.databinding.ActivityMainBinding
+import co.ninetynine.android.exercise.util.Util.getSampleForm
 
-import co.ninetynine.android.exercise.adapters.SectionsListAdapter;
-import co.ninetynine.android.exercise.model.Page;
-import co.ninetynine.android.exercise.util.Util;
-import timber.log.Timber;
+class MainActivity : AppCompatActivity() {
 
-public class MainActivity extends AppCompatActivity {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-  @Override protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    Timber.plant(new Timber.DebugTree());
-    setContentView(R.layout.activity_main);
-
-    Page page = Util.getSampleForm(this);
-    page.logPage();
-    TextView title = findViewById(R.id.title);
-    RecyclerView sections = findViewById(R.id.sections);
-
-    title.setText(page.title);
-    sections.setAdapter(new SectionsListAdapter(page.sections));
-  }
+        getSampleForm(this)?.let { page ->
+            page.logPage()
+            binding.title.text = page.title
+            binding.sections.adapter = SectionsListAdapter(page.sections)
+        }
+    }
 }
